Split out `getThumbnailSizeDimensions`
