+++ client/lib/media/utils/index.js
@@ -11,7 +11,6 @@ const debug = debugFactory( 'calypso:media' );
  */
 import { reduxGetState } from 'lib/redux-bridge';
 import {
-	ThumbnailSizeDimensions,
 	GalleryColumnedTypes,
 	GallerySizeableTypes,
 	GalleryDefaultAttrs,
@@ -44,31 +43,7 @@ export { isSupportedFileTypeForSite } from 'lib/media/utils/is-supported-file-ty
 export { isExceedingSiteMaxUploadSize } from 'lib/media/utils/is-exceeding-site-max-upload-size';
 export { isVideoPressItem } from 'lib/media/utils/is-videopress-item';
 export { playtime } from 'lib/media/utils/playtime';
-
-/**
- * Returns an object containing width and height dimenions in pixels for
- * the thumbnail size, optionally for a given site. If the size cannot be
- * determined or a site is not passed, a fallback default value is used.
- *
- * @param  {string} size Thumbnail size
- * @param  {object} site Site object
- * @returns {object}      Width and height dimensions
- */
-export function getThumbnailSizeDimensions( size, site ) {
-	let width, height;
-
-	if ( site && site.options ) {
-		width = site.options[ 'image_' + size + '_width' ];
-		height = site.options[ 'image_' + size + '_height' ];
-	}
-
-	if ( size in ThumbnailSizeDimensions ) {
-		width = width || ThumbnailSizeDimensions[ size ].width;
-		height = height || ThumbnailSizeDimensions[ size ].height;
-	}
-
-	return { width, height };
-}
+export { getThumbnailSizeDimensions } from 'lib/media/utils/get-thumbnail-size-dimensions';
 
 /**
  * Given an array of media items, returns a gallery shortcode using an
